small fixes and profile dropdown functional on main page
