+++ client/src/index.css
@@ -31,6 +31,8 @@ body {
   border-radius: 30px;
   background-color: black;
   color: white;
+  padding: 15px;
+  font-size: 18px;
   box-shadow: 0px 5px 5px white;  
 }
 
@@ -40,6 +42,8 @@ body {
   color: rgba(255, 255, 255, 0.539);
   border: 2px inset rgb(92, 92, 92);
   box-shadow: 0px 5px 5px white;  
+  padding: 15px;
+  font-size: 18px;
 }
 
 .btn:hover {
