+++ client/src/styles/InputField.css
@@ -1,7 +1,7 @@
 .input-field {
     width: 250px;
     margin: 10px 0px;
-    padding: 7px;
+    padding: 15px;
     font-size: 20px;
     border-radius: 30px;
 }
